remove C++ addon
