--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "native/secp256k1"]
-	path = native/secp256k1
-	url = https://github.com/bitcoin-core/secp256k1.git
