+++ package.json
@@ -1,15 +1,12 @@
 {
   "name": "tiny-secp256k1",
   "version": "1.1.6",
-  "description": "A tiny secp256k1 native/JS wrapper",
+  "description": "A tiny secp256k1 JS",
   "main": "index.js",
-  "gypfile": true,
   "engines": {
     "node": ">=6.0.0"
   },
   "scripts": {
-    "build": "node-gyp rebuild",
-    "install": "npm run build || echo \"secp256k1 bindings compilation fail. Pure JS implementation will be used.\"",
     "unit": "tape tests/index.js | tap-difflet -p",
     "test-browser": "browserify tests/index.js > tests/bundle.js",
     "test": "npm run unit"
@@ -25,18 +22,12 @@
   },
   "homepage": "https://github.com/bitcoinjs/tiny-secp256k1#readme",
   "dependencies": {
-    "bindings": "^1.3.0",
-    "bn.js": "^4.11.8",
     "create-hmac": "^1.1.7",
-    "elliptic": "^6.4.0",
-    "nan": "^2.13.2"
+    "elliptic": "^6.4.0"
   },
   "devDependencies": {
     "browserify": "^16.2.2",
     "tap-difflet": "^0.7.2",
     "tape": "^4.9.0"
-  },
-  "browser": {
-    "./index.js": "./js.js"
   }
 }
